requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
jinja2==3.1.2
python-multipart==0.0.6
python-dotenv==1.0.0
murf==2.0.0
requests==2.31.0
assemblyai==0.43.1
google-generativeai==0.5.4
pymongo==4.6.0
motor==3.3.2
ddgs==9.5.4
aiohttp==3.9.5
feedparser==6.0.10
certifi>=2024.0.0
# Authentication & Security dependencies
bcrypt==3.2.0
pyjwt[crypto]==2.8.0
email-validator==2.1.0
validate_email_address
dnspython==2.4.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
authlib==1.2.0
python-jose[cryptography]==3.3.0
itsdangerous==2.1.2
pyOpenSSL==23.2.0
argon2-cffi==23.1.0
rate-limiter==0.0.3
redis==5.0.1
# Security Headers
secure==0.3.0
--prefer-binary
